SP-359 refactor: 기존 controller 디렉터리 구조 변경에 따른 import 경로 반영
